Moving Files
Moving Vars and Enum to const/ folder
