+++ src/bg/AnimalHandler.js
@@ -1,5 +1,5 @@
 import Counter from "../util/Counter";
-import Vars from "../util/Vars";
+import Vars from "../const/Vars";
 
 export default class AnimalHandler {
