+++ src/const/Dialogue.js
@@ -1,4 +1,4 @@
-import Enum from "../util/Enum";
+import Enum from "../const/Enum";
 
 //  Line break -#
