+++ src/scenes/PlayScene.js
@@ -4,7 +4,7 @@ import Scenery from "../bg/Scenery";
 import MapBuilder from "../bg/MapBuilder";
 import Shadow from "../bg/Shadow";
 import Soldier from "../gameobjects/Soldier";
-import Vars from "../util/Vars";
+import Vars from "../const/Vars";
 import KeyboardMapper from "../util/KeyboardMapper";
 import ControlKeys from "../util/ControlKeys";
 import SpriteController from "../util/SpriteController";
@@ -12,7 +12,7 @@ import MapTracker from "../util/MapTracker";
 import Collectible from "../gameobjects/Collectible";
 import BirdHandler from "../bg/BirdHandler";
 import AnimalHandler from "../bg/AnimalHandler";
-import Enum from "../util/Enum";
+import Enum from "../const/Enum";
 import Rock from "../gameobjects/Rock";
 import SaveData from "../util/SaveData";
 import Tutorial from "../classes/Tutorial";
